pyproject.toml
[project]
name = "education-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "datasets>=4.5.0",
    "google>=3.0.0",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "huggingface-hub>=0.36.1",
    "llama-cpp-python>=0.3.16",
    "openai>=2.17.0",
    "peft>=0.18.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "trl<0.9.0",
    "unsloth>=2025.10.8",
    "xformers>=0.0.34",
]
